Update: Game, Menu, and Jukebox.
Game has been updated to include functions from the collision lab. This
will allow their use if need be, but anything unused will be removed.
Game also had a new variable called c_run added which will be used as an
over all game state variable. If this variable is set to false, then the
game is not running and should be displaying the main menu. Once the
variable is set to true, the game is running and the pause menu should
only ever be displayed when the space key is hit. This means we can
possibly instigate a new game or restart game function at a later time.
This will not effect anyone unless you need to directly call the main or
pause menus. Menu was updated to remove the instance usage. Simply call
the menu function you need. The all files had their code cleaned up.
This process of cleaning the code in on going and will be updated as the
code is formatted and added.