Added temporary main menu image
Added a temporary mainmenu bmp to be used for testing. The pause and
main menu need to be visually different to ensure that the right menu is
being displayed at the right time. This allows us to see this early on
before the final version of the main menu is decided upon.