Update: Game
Updated Game.h and .cpp to be more organized. Added the MapHandler code
to create it. Added more comments. Added the updateTile function to
update the background texture variable. Sectioned off groups of
functions within the .cpp to make it easier to find a function (if you
know the sectioning). Also made small changes here and there that don't
come to mind at the moment.