Update: Game And Menu
Updated the Game code to include some of the object instance update
calls. Changed the code which handles the pausing of the game in
Game.cpp. Changed the Menu.h file to be more reliant on the config.h
file. Changed the Menu.cpp code to take into account the menus folder.
Menus folder added to clean up the menu images into a single location
without taking up space. This also allows for the creation of more types
of menus or prototypes. Removed unnecessary code from Game.cpp. Added
more code to Game.cpp required by other members of the team and files
within the project.