Update: Game and config.h
Updated Game.cpp and .h to remove junk code, alter calls to implement
the proper singleton code. Updated the key() function to use different
keys recommended by the team. Prepared the calls to swap items and
weapons. Prepared calls for item and weapon use. Fixed the starting
texture call to use a variable defined in teh config.h. Added a new
variable to config.h called START_TILE which stores the starting tile
required by the Game.cpp init function. Updated the config.h screen
settings to match the image size. Corrected Game.h to comment out unused
code at this time. This code in the .h and the code in the .cpp which is
unused will be removed near the end.