Update: Game, Menu, and Jukebox.
Updated Game to have a method to prevent the player from moving or
performing an action while the game is paused. This change needs
testing. Changed menu to include a win/lose function to display the win
or lose menu depending on the required display. Changed jukebox to have
a dedicated background song play. Also added to the config.h file to
allow easy alterations to the various menu images and sounds if their
names change. Config.h needs to have the pickup, hurt, attack and any
other sounds defined in them so that I may create a function dedicated
to their sounds.