Minor update: Game and Menu.
Added the variables that will store if the game is over, and if the win
condition was met. This iv c_gameOver is true, then the game is over and
the c_winCondition should be checked so that the proper win/lose screen
will be displayed. Also made a small spelling correction in menu.