Fixed Seg Fault.
Moved the texture loading into menu.init(). This allows the preloading
of the menu textures. Removed the texture loading code form the menu
constructor. Commented out the win/lose screen prototypes. So far this
should allow the game to run, but the pause menu does not want to
display correctly. I am currently trying to debug this and it appears as
if the file is corrupt in the game, yet outside of the game it appears
fine. I am also trying to fix the centering of the images as it appears
that my calculations were not correct... having alot of trouble trying
to center the dam things.