Update: Game, Menu, and Jukebox. See description!!!!!
Updated Game.cpp and .h to include the ability for us to hit escape and
then enter to quit while the game is paused. No confirmation message is
displayed between the escape key and the enter key yet, however that is
planned feature. Updated Menu to prepare the class to contain the quit
confirmation screen. Updated Jukebox to have specific sound functions.
Thus if you want to call the attack sound, simply call the designated
function from the Jukebox class. Also included more definitions for the
config.h file. These include the quit confirmation screen definition,
and all of the sound definitions. These changes besides the new quit
ability, should not effect the game in anyway. I have yet to test these
due to being kicked out of the computer lab and having no remote way to
test this. I will test these asap an issue a fix for anything that gets
broken.