Update: Images and Config.h
Updated the images for the menus. Changed some of the menus and added in
some confirmation dialogs. This is preperation for a future change that
is currently waiting to be tested. I only updated the images first to
allow them to be tested first before the code changes are implemented.
Config.h had some updates done to it adding in a few more textures.